requirements.txt
absl-py
astunparse
cachetools
certifi
charset-normalizer
cycler
flatbuffers
fonttools
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
keras
Keras-Preprocessing
kiwisolver
libclang
Markdown
MarkupSafe
matplotlib
numpy
oauthlib
opencv-python
opt-einsum
packaging
Pillow
protobuf
pyasn1
pyasn1-modules
pyparsing
python-dateutil
requests
requests-oauthlib
rsa
six
termcolor
toml
toolz
torch
torchvision
tornado
typing_extensions
urllib3
Werkzeug
wrapt
